+++ packages/typespec-python/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release
 
+## 0.63.2
+
+- Bump @typespec/http-client-python to 0.33.0
+
+## 0.63.1
+
+- Bump @typespec/http-client-python to 0.30.1
+
 ## 0.63.0
 
 - [#4503](https://github.com/Azure/typespec-azure/pull/4503) Bump @typespec/http-client-python to 0.30.0